refactor: barrel files in copy and paste plugin
