+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/copy_and_paste/commands/copy_and_paste_commands.dart
@@ -1,0 +1,8 @@
+import 'package:appflowy/plugins/document/presentation/editor_plugins/copy_and_paste/commands/commands.dart';
+import 'package:appflowy_editor/appflowy_editor.dart';
+
+final List<CommandShortcutEvent> copyAndPasteCommands = [
+  customCopyCommand,
+  customPasteCommand,
+  customCutCommand,
+];
